AGDROID-686 Move to org.aerogear.android.ags.auth package
