+++ auth/src/main/java/org/aerogear/android/ags/auth/AuthServiceConfig.java
@@ -1,0 +1,8 @@
+package org.aerogear.android.ags.auth;
+
+/**
+ * Authentication service configuration
+ */
+public class AuthServiceConfig {
+    // TODO: to be implemented.
+}
