+++ auth/src/main/java/org/aerogear/android/ags/auth/ClientRole.java
@@ -1,0 +1,7 @@
+package org.aerogear.android.ags.auth;
+
+public final class ClientRole extends AbstractRole {
+    public ClientRole(final String roleName) {
+        super(roleName);
+    }
+}
