+++ auth/src/main/java/org/aerogear/android/ags/auth/IRole.java
@@ -1,0 +1,5 @@
+package org.aerogear.android.ags.auth;
+
+public interface IRole {
+    String getRoleName();
+}
