+++ auth/src/main/java/org/aerogear/android/ags/auth/RealmRole.java
@@ -1,0 +1,7 @@
+package org.aerogear.android.ags.auth;
+
+public final class RealmRole extends AbstractRole {
+    public RealmRole(final String roleName) {
+        super(roleName);
+    }
+}
