+++ auth/src/main/java/org/aerogear/android/ags/auth/credentials/IIntegrityCheckParameters.java
@@ -1,0 +1,10 @@
+package org.aerogear.android.ags.auth.credentials;
+
+public interface IIntegrityCheckParameters {
+
+    String getAudience();
+    String getIssuer();
+    String getPublicKey();
+    boolean isValid();
+    String serialize();
+}
