+++ auth/src/main/java/org/aerogear/android/ags/auth/impl/OIDCAuthCodeImpl.java
@@ -1,0 +1,25 @@
+package org.aerogear.android.ags.auth.impl;
+
+import org.aerogear.android.ags.auth.credentials.ICredential;
+import org.aerogear.mobile.core.configuration.ServiceConfiguration;
+
+import java.security.Principal;
+
+/**
+ * Authenticates the user by using OpenID Connect.
+ */
+public class OIDCAuthCodeImpl extends OIDCTokenAuthenticatorImpl {
+
+    public OIDCAuthCodeImpl(final ServiceConfiguration serviceConfig) {
+        super(serviceConfig);
+    }
+
+    /**
+     * @param credential Ignored.
+     * @return authenticated Principal
+     */
+    @Override
+    public Principal authenticate(final ICredential credential) {
+        throw new IllegalStateException("Not implemented");
+    }
+}
