--- auth/src/main/java/org/aerogear/auth/AuthServiceConfig.java
@@ -1,8 +1,0 @@
-package org.aerogear.auth;
-
-/**
- * Authentication service configuration
- */
-public class AuthServiceConfig {
-    // TODO: to be implemented.
-}
