--- auth/src/main/java/org/aerogear/auth/ClientRole.java
@@ -1,7 +1,0 @@
-package org.aerogear.auth;
-
-public final class ClientRole extends AbstractRole {
-    public ClientRole(final String roleName) {
-        super(roleName);
-    }
-}
