--- auth/src/main/java/org/aerogear/auth/IRole.java
@@ -1,5 +1,0 @@
-package org.aerogear.auth;
-
-public interface IRole {
-    String getRoleName();
-}
