--- auth/src/main/java/org/aerogear/auth/RealmRole.java
@@ -1,7 +1,0 @@
-package org.aerogear.auth;
-
-public final class RealmRole extends AbstractRole {
-    public RealmRole(final String roleName) {
-        super(roleName);
-    }
-}
