--- auth/src/main/java/org/aerogear/auth/credentials/ICredential.java
@@ -1,8 +1,0 @@
-package org.aerogear.auth.credentials;
-
-/**
- * Base interface for credential objects.
- */
-public interface ICredential {
-
-}
