--- auth/src/main/java/org/aerogear/auth/credentials/IIntegrityCheckParameters.java
@@ -1,12 +1,0 @@
-package org.aerogear.auth.credentials;
-
-import org.json.JSONException;
-
-public interface IIntegrityCheckParameters {
-
-    String getAudience();
-    String getIssuer();
-    String getPublicKey();
-    boolean isValid();
-    String serialize() throws JSONException;
-}
